fix: sidebar item refactored
